총알 동기화 일부 완료

+++ Client/Client/Player.cpp
@@ -86,12 +86,12 @@ bool Player::GetConnected()
 	return m_connected;
 }
 
-int Player::GetX()
+float Player::GetX()
 {
 	return m_PositionX;
 }
 
-int Player::GetY()
+float Player::GetY()
 {
 	return m_PositionY;
 }
